+++ ersatz/src/main/java/io/github/cjstehno/ersatz/cfg/RequestWithContent.java
@@ -17,11 +17,13 @@ package io.github.cjstehno.ersatz.cfg;
 
 import io.github.cjstehno.ersatz.encdec.DecodingContext;
 import io.github.cjstehno.ersatz.match.BodyMatcher;
+import io.github.cjstehno.ersatz.match.BodyParamMatcher;
 import org.hamcrest.Matcher;
 
 import java.util.function.BiFunction;
 
 import static io.github.cjstehno.ersatz.match.BodyMatcher.bodyMatching;
+import static io.github.cjstehno.ersatz.match.BodyParamMatcher.bodyParamMatching;
 import static org.hamcrest.Matchers.equalTo;
 
 /**
@@ -93,9 +95,9 @@ public interface RequestWithContent extends Request {
     RequestWithContent decoder(final String contentType, final BiFunction<byte[], DecodingContext, Object> decoder);
 
     /**
-     * Specifies a custom body content converter function. The function will have the client request body content as a byte array and it will be
-     * converted into the specified output type. Generally the conversion is used when comparing the client request with the configured request
-     * body expectation.
+     * Specifies a custom body content converter function. The function will have the client request body content as a
+     * byte array and it will be converted into the specified output type. Generally the conversion is used when
+     * comparing the client request with the configured request body expectation.
      *
      * @param contentType the content type that the convert will handle
      * @param decoder     the conversion function
@@ -106,32 +108,45 @@ public interface RequestWithContent extends Request {
     }
 
     /**
-     * Configures an expectation matching parameters contained in the request body. The specified value must exist in the list
-     * of parameters for the provided key.
+     * Configures an expectation matching parameters contained in the request body. The specified value must exist in
+     * the list of parameters for the provided key.
      *
-     * @param name  the parameter name
+     * @param name  the request body parameter name
      * @param value the expected parameter value
      * @return a reference to this request
      */
-    RequestWithContent param(final String name, final String value);
+    default RequestWithContent param(final String name, final String value) {
+        return param(bodyParamMatching(name, value));
+    }
 
     /**
-     * Configures an expectation matching parameters contained in the request body. The specified values must exist in the list
-     * of parameters for the provided key.
+     * Configures an expectation matching parameters contained in the request body. The specified values must exist in
+     * the list of parameters for the provided key.
      *
-     * @param name   the parameter name
+     * @param name   the request body parameter name
      * @param values the expected parameter values
      * @return a reference to this request
      */
-    RequestWithContent param(final String name, final Iterable<? super String> values);
+    default RequestWithContent param(final String name, final Iterable<? super String> values) {
+        return param(bodyParamMatching(name, values));
+    }
 
     /**
-     * Configures an expectation matching parameters contained in the request body. The specified matchers must be satisfied
-     * by the parameters mapped to the provided named parameter.
+     * Configures an expectation matching parameters contained in the request body. The specified matchers must be
+     * satisfied by the parameters mapped to the provided named parameter.
      *
-     * @param name     the parameter name
-     * @param matchers the expected parameter value matchers
+     * @param name    the request body parameter name
+     * @param matcher the matcher for the expected parameter values
      * @return a reference to this request
      */
-    RequestWithContent param(final String name, final Matcher<Iterable<? super String>> matchers);
+    default RequestWithContent param(final String name, final Matcher<Iterable<? super String>> matcher) {
+        return param(bodyParamMatching(name, matcher));
+    }
+
+    /**
+     * Configures an expectation matching request body parameters.
+     *
+     * @param bodyParamMatcher the request body parameter matcher
+     */
+    RequestWithContent param(final BodyParamMatcher bodyParamMatcher);
 }