+++ ersatz/src/main/java/io/github/cjstehno/ersatz/impl/ErsatzRequest.java
@@ -19,6 +19,7 @@ import io.github.cjstehno.ersatz.cfg.HttpMethod;
 import io.github.cjstehno.ersatz.cfg.Request;
 import io.github.cjstehno.ersatz.cfg.Response;
 import io.github.cjstehno.ersatz.encdec.ResponseEncoders;
+import io.github.cjstehno.ersatz.impl.matchers.HttpMethodMatcher;
 import io.github.cjstehno.ersatz.impl.matchers.RequestSchemeMatcher;
 import io.github.cjstehno.ersatz.match.HeaderMatcher;
 import io.github.cjstehno.ersatz.match.PathMatcher;
@@ -61,7 +62,7 @@ public class ErsatzRequest implements Request {
      * @param noResponse     whether or not this is a request with an empty response (defaults to false)
      */
     public ErsatzRequest(final HttpMethod meth, final PathMatcher pathMatcher, final ResponseEncoders globalEncoders, final boolean noResponse) {
-        matchers.add(RequestMatcher.method(meth == ANY ? isOneOf(GET, HEAD, POST, PUT, DELETE, PATCH, OPTIONS, TRACE) : equalTo(meth)));
+        matchers.add(new HttpMethodMatcher(meth));
         matchers.add(pathMatcher);
 
         this.globalEncoders = globalEncoders;