+++ ersatz/src/main/java/io/github/cjstehno/ersatz/impl/matchers/HttpMethodMatcher.java
@@ -1,0 +1,41 @@
+/**
+ * Copyright (C) 2022 Christopher J. Stehno
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.github.cjstehno.ersatz.impl.matchers;
+
+import io.github.cjstehno.ersatz.cfg.HttpMethod;
+import io.github.cjstehno.ersatz.server.ClientRequest;
+import lombok.RequiredArgsConstructor;
+import org.hamcrest.BaseMatcher;
+import org.hamcrest.Description;
+
+import static io.github.cjstehno.ersatz.cfg.HttpMethod.ANY;
+
+/**
+ * A matcher used to match the HTTP method of a request.
+ */
+@RequiredArgsConstructor
+public class HttpMethodMatcher extends BaseMatcher<ClientRequest> {
+
+    private final HttpMethod method;
+
+    @Override public boolean matches(final Object actual) {
+        return method == ANY || method == ((ClientRequest) actual).getMethod();
+    }
+
+    @Override public void describeTo(final Description description) {
+        description.appendText("HTTP method is " + method);
+    }
+}