+++ CHANGES.md
@@ -3,3 +3,5 @@
 * #150: Reduce the number of threads configured
 * #141: Update commons-io
 * #151: Fixed bug related to query param matching
+* Added predicate function matchers
+* Refactored the matcher system from the API interface down